pyproject.toml
[project]
name = "strava-upload"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Max F", email = "max.f@hmne.com" }
]
dependencies = [
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0.post1",
    "stravalib>=1.4",
    "typer>=0.9.0",
    "loguru>=0.7.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
upload = "strava_upload:main.upload"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.rye.scripts]
devserver = "uvicorn strava_upload.auth:app --reload"
auth = "uvicorn strava_upload.auth:app"

[tool.hatch.metadata]
allow-direct-references = true
